server/requirements.txt
# server/requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
asyncpg==0.29.0
pydantic==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
alembic==1.13.3
python-dotenv==1.0.1